azure: populate sandbox vm name earlier
Fixes: #275

There might be a state in which the creation of a VM has been triggered
but the respective api call hasn't returned yet. If a create-vm call is
then cancelled at the call site due to a timeout, the VM resources will
not be garbage collected, because the sandbox's vm name has not been
populated yet.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>