webhook: add automated tests and runner script
This added three Bats tests for the webhook:
 - test it can mutate a pod
 - test it should not mutate non-peerpods
 - test default parameters can be changed

Being the last one skipped because it is not passing.

It is also introduced a runner script (run-local.sh) which will
bootstrap the test environment with Kind and afterwards run those tests.
At the end of the execution the cluster and created resources are
deleted, however, you can retain them by running the script in debug
mode:

 $ ./tests/e2e/run-local.sh -d

However, the recommend way to run the e2e is with `make`:

$ make test-e2e

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>