libvirt: separate out the ubuntu packer template into a separate dir
Move ubuntu target to its own separate folder so that
we can keep the same call to packer build when we introduce
rhel. The IMAGE_FILE target in Makefile is also slightly changed so that it
can be reused when we introduce the rhel image build.

Packer limitations: Another reason for doing it this way is that packer
does not give an easy way to have two separate build scripts in the same
dir. We can either add all builders to the same script file or as we do here,
separate out the builds into their own dirs. It does add some code duplication,
specifically the variable definitions.

No functional change.

Fixes: #384
Signed-off-by: Bandan Das <bsd@redhat.com>