test/e2e: allow to change the overlay pre-apply
This introduces a mechanism where install overlay files (install/overlays/*/*)
can be modified before the overlay is applied.

It was created the InstallOverlay interface that defines three operations:
apply, delete, and edit. The former two are common accross cloud
providers so a default implementation is made. The edit operator is
specific for each provider so only a stub implementation is made.

The InstallOverlay.Edit() is called before Apply() in
CloudAPIAdaptor.Deploy().

Fixes #592
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>