peerpodconfig-ctrl: Handle NodeSelectors
peerpodconfig-ctrl creates the cloud-api-adaptor daemonset and
in order to ensure the pod gets created on the right nodes, it
uses nodeselector in the pod spec. The nodeSelector in the pod spec
is a map[string]string so it makes sense to use the same type
in our api spec. This patch makes the relevant changes to handle
nodeSelector for the CAA pod

Fixes: #853

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>