test/libvirt: allow libvirt conn URI to be specified
The libvirt tests currently assumes that it is always running on a
local instance and hard codes "qemu:///system" as the URI to create the
Connection object with.

This allows us to specify libvirt_conn_uri in the libvirt properties
file to change it to a different instance.

Also updates test/provisioner/e2e/README.md to add this option to the
documentation and clarify the difference with libvirt_uri.

Fixes: #1164

Signed-off-by: Derek Lee <derlee@redhat.com>