workflow/caa_build_and_push: new tags inputs
Changed the caa_build_and_push.yaml workflow to receive new inputs:

* dev_tags: Comma-separated list of tags for the dev built image (e.g. latest,ci-dev). By default uses the values from hack/build.sh
* release_tags: Likewise but for the release built image

This will allow us to publish cloud-api-adaptor images with tags
differently from the defaults. Examples: a) for the sake of CI we don't need to
publish the release builds with `latest` tag, we may also record the pull
request number on the tag b) a release workflow may want to tag the image
with a git tag.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>