libvirt: poll the domain to get IPs
Instead of sleep for a fixed amount of time (30s) before try to
get the domain IP addresses, now it will be polling the domain every 3
seconds and fail if it is unable to get the IPs after 20 retries.

Note that this increased the maximum wait time to ~60s because on CI the
podvm is getting more than 30s to be ready.

Fixes #1216
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>