test/e2e: implement on-prem cluster for AWS
Created an `Cluster` interface with generic operations to create/delete
a kubernetes cluster as well as return the kubeconfig file. Then
implemented an on-prem cluster, which is essentially a cluster already
deployed and running.

On-prem cluster is the default for the AWS provisioner. The
implementation will look for the kubeconfig file on the default
locations (KUBECONFIG variable or ~/.kube/config).

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>