workflows/podvm_builder: read properties from version.yaml
The podvm_builder workflow currently have hard-coded the versions to
Golang and Rust, among other properties. On updating any of those
properties the workflow code should be changed, and this break the
rationale of having the versions.yaml as a centralized source of truth
for the project dependencies. Therefore, let's change the workflow to
read the properties from versions.yaml instead.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>