workflows: convert to callable builder/binaries/podvm
The podvm_builder, podvm_binaries, and podvm workflows were converted to
`workflow_call` so that they can be re-used.

Before, on release time, `podvm_builder` were first triggered on a `release` event
then the others followed upon finish of the previous. With this change,
the `podvm_on_release` workflow was created to be the entry point for
the `release` event, so the (now) callable `podvm_builder`, `podvm_binaries`, and
`podvm` are directly called.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>