workflows/podvm: event-based image tagging
The current behavior of builder/binaries/podvm workflows is to push
images with the `latest` and commit SHA tags. Now that those workflows
are callable there is a need to tag those images differently depending
on the use case. The approach took here was to used different tags based
on the event that triggered the workflow:

* on `release` event, tag with `latest` and the commit SHA. Default
  behavior does not change
* on `pull_request` or `pull_request_target` uses the `ci-prN` where `N`
  is the pull request number. Does not make sense to push `latest` or
commit SHA to an image that is transient.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>