podvm: reset the CAA branch if it exists
When it tries to create the `CAA_SRC_REF` branch and it exists already, it
will reset the branch instead of simply fail the checkout operation.

That error can happen, for example, if `CAA_SRC_REF` is passed to the builder
image then that image is used for the binaries build and you passed
the same `CAA_SRC_REF`, i.e., on the builder image the repository is
already branched to `CAA_SRC_REF` and in the binary image build it will
try to create the branch same again (so failing). The checkout with `-B`
will reset the branch instead of fail.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>