podvm: allow to inject a custom agent-config.toml
Introduced the STATIC_AGENT_CONFIG variable that should be used to inject
a custom agent-config.toml in the podvm image. The value of STATIC_AGENT_CONFIG
should be a relative path to "podvm/file", for example, STATIC_AGENT_CONFIG=/etc/agent-config.toml
and podvm/file/etc/agent-config.toml must exist.

It will include $STATIC_AGENT_CONFIG in the image and /etc/systemd/system/kata-agent.service
is tweaked to pass $STATIC_AGENT_CONFIG as the kata-agent configuration. As a result (1) the
/run/peerpod/agent-config.toml is created but disregarded, (2) auth-json-secret secret is also
disregarded.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>