docker: Fix userdata bind mount path inside the container
For docker provider we use systemd container which mounts /run
as a tmpfs volume.
This mounting of /run happens after docker has bind mounted the
userdata file inside the container and consequently the userdata file is no
longer accessible.

This commit changes the path of the userdata file to not use `/run` for
the docker provider.

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>