workflows: add resources removal step to AWS
The e2e framework can get aborted in the middle of execution, for
example, when a test case don't catch an error properly. Depending on
the timing it was interrupted it will leave resources in AWS because the
teardown code won't run. This introduced a new script and workflow's
step to run clean up routines at end of the workflow.

Note 1:

Currently the e2e testing for AWS isn't prepared to run on parallel jobs on
same account because it doesn't create unique resources names. So this clean
up script can't run in parallel jobs either. It will work out fine on
nightly CI jobs though.

Note 2:

In the future we might replace this by a AWS-side solution

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>