test/provision: use non-default AWS vmimport role
The vmimport role is used to import an EBS snapshot in the process to
create an AMI image of the podvm. Until now it has used the default
'vmimport' role name, but src/cloud-api-adaptor/aws/raw-to-ami.sh may
override vmimport's policy so that the S3 bucket used on e2e test has
access denied, hence, the importing of EBS snapshot fails. Instead,
let's used an new vmimport role only for the e2e tests.

Note : that role doesn't need to be unique per execution because the
S3 bucket is already shared among e3e tests executions.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>