test/provision: use eksctl to create EKS
The current support to create EKS relies on using the AWS SDK for Go
libraries. This has made the implementation a bit complex but it works,
however, in following changes it will be switched to use Ubuntu workers
which would require more uneeded code as we could be using a tool like
eksctl instead. That's exactly what this commit does: use eksctl to
create the EKS cluster.

Updated to k8s 1.34 as 1.26 is deprecated. Don't need to handle roles,
CNI plug-in, etc, because it all carried out by eksctl. However, still
relying on the VPC and subnets already created to run the podvms.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>