workflows/e2e_aws: install qemu-img in the runner
qemu-img is used to convert the podvm disk from qcow2 to raw to then
upload to AWS S3, so it is a requirement.

In the current onprem (kcli) job it's installed via
src/cloud-api-adaptor/libvirt/config_libvirt.sh as a collateral of
installing qemu-kvm. But with EKS job it doesn't run that script so
let's install the tool in a workflow step.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>